fix: (#45) hours worked changed from int to float

+++ noscrum/db.py
@@ -29,7 +29,7 @@ class Work(db.Model):
     __tablename__ = "work"
     id = sa.Column(sa.Integer(), primary_key=True)
     work_date = sa.Column(sa.Date(), nullable=False)
-    hours_worked = sa.Column(sa.Integer(), nullable=False)
+    hours_worked = sa.Column(sa.Float(), nullable=False)
     task_id = sa.Column(sa.Integer(), sa.ForeignKey("task.id"))
     status = sa.Column(sa.String(12), nullable=True, default="To-Do")
     user_id = sa.Column(sa.Integer(), sa.ForeignKey("user.id"))
